added dtos and jpqls

+++ library-project/src/main/java/com/capgemini/library_project/entities/Book.java
@@ -3,7 +3,6 @@ package com.capgemini.library_project.entities;
 import java.util.List;
 
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonBackReference;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 
@@ -21,6 +20,7 @@ import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 @Entity
 @Data
 @AllArgsConstructor
@@ -41,11 +41,11 @@ public class Book {
 
     @NotNull(message = "Total Copies is required")
     @Positive(message = "Total Copies cannot be negative")
-    private Integer totalCopies;
+    private Long totalCopies;
 
     @NotNull(message = "Available Copies is required")
     @Positive(message = "Available Copies cannot be negative")
-    private Integer availableCopies;
+    private Long availableCopies;
 
     private String bookCover;
